Committing the changes made by Frédéric Vivien before the weekend. His summary (edited MV):
- Correction of bug in xf_draw_polygon.

- Now setlinestyle handle style_val, cap_style, and join_style (passed
as pointers). All the calls to setlinestyle are updated and cleaned
most (all ?) of the xf_draw_(something using a polyline).

- The META_STRETCHDIB is now processed by wmftofig. This is one of the
four functions handling bitmaps in wmf... and the one I have examples
of use.

- cleaned up the META_STRETCHDIB code, for example by no more
computing 5 times the same ratio, and ... to avoid rounding-up.

And finally: the problem file cell.wmf! Frédéric gave permission
to include it.