configure: when explicit directory for libxml2 is not given, use pkg-config
xml2-config is deprecated upstream. Arguably, it should never have used
manual invocation of xml2-config to begin with, since legacy xml2-config
usage was expected to be activated via AM_PATH_XML2 (which today warns
it is obsolete and says to use PKG_CHECK_MODULES directly, and does
indeed now use PKG_CHECK_MODULES itself).

So we migrate to the recommended alternative. When --with-libxml2=DIR is
passed with a directory value, we fall back to xml2-config. This usage
is frankly unusual and pkg-config has better ways of specifying
alternative installs of a library (use PKG_CONFIG_PATH=DIR/lib/pkgconfig
for that) but to avoid making too many controversial changes, don't
touch that logic.

Migrating to pkg-config fixes overlinking to libxml2's dependencies, in
particular random ICU components which then cause libwmf to be fatally
broken since it links to an incorrect, different version of ICU. This
causes issues for example when rebuilding with a new ICU version, and
libxml2 is rebuilt since it directly requires ICU, but libwmf doesn't
officially require it and skips the rebuild. (And it is entirely up to
chance whether libwmf is accidentally linked to ICU anyway, and
therefore breaks.)